FEAT : 검수내역 불러오기 & 결제요청서 리스트 불러오기 구현

+++ src/main/java/com/meta/overwash/domain/ReceiptDTO.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 @NoArgsConstructor
 @ToString
 public class ReceiptDTO {
-	private Integer receiptId;
+	private Long receiptId;
 	private String paymentMethod;
 	private PaymentRequestDTO pr;
 }
